fix evolve address derivation, server metrics, and dashboard
- Rewrite evolve-utils.ts with correct keccak256-based address derivation
  (accountIdToAddress, addressToAccountId, buildTransferData with 32-byte AccountIds)
- Fix server event metrics: use monotonic counters instead of recounting
  from the truncated event log (root cause of total requests halving at ~320)
- Simplify token discovery in pool.ts: deterministic derivation from
  genesis AccountId 65535 instead of trial-and-error
- Dashboard: show server-side total requests counter, remove bouncy
  block-level metric cards, add chain/WS connection indicators
- Set treasury to faucet address for token recirculation
- Bump Dockerfiles to bun:1-alpine

+++ simulator/src/agent.ts
@@ -245,8 +245,10 @@ export class Agent {
       });
     }
 
+    // payTo is an Ethereum address — derive the EOA AccountId (32 bytes)
     const payToAccountId = addressToAccountId(to);
     const data = buildTransferData(payToAccountId, amount);
+    // tokenAccountId is already a 32-byte Uint8Array — derive its contract address
     const tokenAddress = accountIdToAddress(this.config.tokenAccountId);
 
     // Claim nonce synchronously before any await to prevent races between concurrent workers
@@ -300,7 +302,7 @@ export function createAgentConfig(
   privateKey: Hex,
   requestsPerSecond: number,
   endpoints: WeightedEndpoint[],
-  tokenAccountId: bigint,
+  tokenAccountId: Uint8Array,
   chainId: number,
 ): AgentConfig {
   const account = privateKeyToAccount(privateKey);
